ui: Add searchbar on subproject-level
The searchbar takes and inputstring and use it as Regex to
search each Column for all subproject titles and status.
Unmatched columns are hidden, matched columns are shown
and the matched strings are highlighted.
The searchbar-component has placeholder-texts in every language,
e.g. "Search projects" or "Search subprojects"
The searchbar-component is used on project- and
subproject-level.
The searchbar inputfield on subproject-level is open by default
and cannot be closed.

Resolves: #422